+++ src/index.ts
@@ -3,5 +3,6 @@ export * from './divide';
 export * from './mean';
 export * from './meanBy';
 export * from './multiply';
+export * from './subtract';
 export * from './sum';
 export * from './sumBy';